# Assignment1/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"scopes": {
					"member": "member",
					"admin": "admin"
				}
			}
		}
	},
	"info": {
		"title": "cse187-assignment1",
		"version": "1.0.0",
		"description": "CSE187 Assignment 1",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {
			"name": "David Harrison ",
			"email": "dcharris@ucsc.edu"
		}
	},
	"paths": {
...
		"/post": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"image": {
											"type": "string"
										},
										"content": {
											"type": "string"
										},
										"posted": {
											"type": "string",
											"format": "date-time"
										},
										"member": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"image",
										"content",
										"posted",
										"member",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"member"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"image": {
										"type": "string"
									},
									"content": {
										"type": "string"
									}
								},
								"required": [
									"image",
									"content"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"image": {
												"type": "string"
											},
											"content": {
												"type": "string"
											},
											"posted": {
												"type": "string",
												"format": "date-time"
											},
											"member": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"image",
											"content",
											"posted",
											"member",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"member"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v0"
		}
	]
}